Refactor into separate files
Split `App` into `Board`, `Solver` and `SudokuSolver`.
