+++ .github/workflows/build.yml
@@ -1,9 +1,6 @@
 name: build
 
-on:
-  pull_request:
-  push:
-    branches: [ master ]
+on: [push, pull_request]
 
 jobs:
   build:
